backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysql-connector-python
sentence-transformers
pdfminer.six
requests
bertopic
numpy